double n'existe pas dans SQL, j'ai change en FLOAT

+++ BD6/requetes.sql
@@ -112,7 +112,7 @@ CREATE TABLE location(
 id_location SERIAL PRIMARY KEY, 
 date_debut_location DATE NOT NULL,
 date_fin_location DATE NOT NULL, 
-montant_total DOUBLE NOT NULL CHECK(montant_total>0),
+montant_total FLOAT NOT NULL CHECK(montant_total>0),
 CHECK ( date_fin_location >= date_debut_location )
 );
